move explosion_anti_cheat_bypass to its own file
